Expose abstract class instead of implementation
