+++ audiorecorder/src/main/java/org/odk/collect/audiorecorder/DaggerSetup.kt
@@ -10,10 +10,9 @@ import dagger.Provides
 import org.odk.collect.audiorecorder.recorder.MediaRecorderRecorder
 import org.odk.collect.audiorecorder.recorder.RealMediaRecorderWrapper
 import org.odk.collect.audiorecorder.recorder.Recorder
-import org.odk.collect.audiorecorder.recording.AudioRecorderService
-import org.odk.collect.audiorecorder.recording.AudioRecorderViewModel
-import org.odk.collect.audiorecorder.recording.RecordingRepository
-import java.lang.IllegalStateException
+import org.odk.collect.audiorecorder.recording.AudioRecorderViewModelFactory
+import org.odk.collect.audiorecorder.recording.internal.AudioRecorderService
+import org.odk.collect.audiorecorder.recording.internal.RecordingRepository
 import javax.inject.Singleton
 
 private var _component: AudioRecorderDependencyComponent? = null
@@ -54,7 +53,7 @@ internal interface AudioRecorderDependencyComponent {
 
     fun inject(activity: AudioRecorderService)
 
-    fun inject(activity: AudioRecorderViewModel.Factory)
+    fun inject(activity: AudioRecorderViewModelFactory)
 }
 
 @Module
